LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-blocking-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL Block Trusted Type policy creation by event listener. Can't find variable: trustedTypes
FAIL Trusted Type policy creation. Can't find variable: trustedTypes
FAIL Block only default Trusted Type policy creation. Can't find variable: trustedTypes
FAIL Policy creation called before addEventListener function will not reached the listener. Can't find variable: trustedTypes
